+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - 0.2.3:
   - **Update** previous line color
+  - **Fix** travel status compared to today
 - 0.2.2:
   - **Fix** zoom to target top limit
 - 0.2.1: